Web build automatization - Linux stuff.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1317 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: e6b64f879b3bfb45d8b350fb15be6672c51ae115

+++ BuildScripts/android.sh
@@ -1,13 +1,31 @@
 #!/bin/bash
 
 # Usage:
-# export FO_SOURCE=<source> && export ANDROID_NDK=<ndk> && $FO_SOURCE/BuildScripts/android.sh
+# export FO_SOURCE=<source> && sudo -E "$FO_SOURCE/BuildScripts/android.sh"
+
+export ANDROID_NDK_VERSION="android-ndk-r12b"
+
+if [ "$EUID" -ne 0 ]; then
+	echo "Run as root"
+	exit
+fi
+
+sudo apt-get -y update
+sudo apt-get -y install build-essential
+sudo apt-get -y install cmake
 
 if [ "$FO_CLEAR" = "TRUE" ]; then
 	rm -rf android
 fi
 mkdir android
 cd android
+
+if [ ! -f "$ANDROID_NDK_VERSION-linux-x86_64.zip" ]; then
+	wget "https://dl.google.com/android/repository/$ANDROID_NDK_VERSION-linux-x86_64.zip" 
+	sudo apt-get -y install unzip
+	unzip "$ANDROID_NDK_VERSION-linux-x86_64.zip" -d "./"
+fi
+export ANDROID_NDK="$PWD/$ANDROID_NDK_VERSION"
 
 mkdir android-project
 cp -r "$FO_SOURCE/BuildScripts/android-project" "./"
